Added working movement over network.
Also started work on handling player joining.

+++ src/main/java/us/aaronweiss/pixalia/server/packets/HandshakePacket.java
@@ -57,7 +57,7 @@ public class HandshakePacket extends Packet {
 	public String virtualHost() {
 		this.ready();
 		if (this.packetType.is(PacketType.INBOUND) && this.hasVhost) {
-			return Utils.readString(this.buffer.readInt(), this.buffer);
+			return Utils.readString(this.buffer.readByte(), this.buffer);
 		}
 		return null;
 	}
